Added FOV changer. Fixed window mode issues. Fixed warnings.

+++ blackshades/Source/Frustum.cpp
@@ -78,7 +78,7 @@ void FRUSTUM::
 	frustum[5][3] = clip[15] + clip[14];
 
 	/* normalize the right plane */
-	t = sqrt(frustum[0][0]*frustum[0][0] 
+	t = sqrtf(frustum[0][0]*frustum[0][0] 
 		+ frustum[0][1]*frustum[0][1] 
 		+ frustum[0][2]*frustum[0][2]);
 	frustum[0][0] /= t;
@@ -93,7 +93,7 @@ void FRUSTUM::
 	frustum[1][3] = clip[15] + clip[12];
 
 	/* normalize the left plane */
-	t = sqrt(frustum[1][0]*frustum[1][0] 
+	t = sqrtf(frustum[1][0]*frustum[1][0] 
 		+ frustum[1][1]*frustum[1][1] 
 		+ frustum[1][2]*frustum[1][2]);
 	frustum[1][0] /= t;
@@ -108,7 +108,7 @@ void FRUSTUM::
 	frustum[2][3] = clip[15] + clip[13];
 
 	/* normalize the bottom plane */
-	t = sqrt(frustum[2][0]*frustum[2][0] 
+	t = sqrtf(frustum[2][0]*frustum[2][0] 
 		+ frustum[2][1]*frustum[2][1] 
 		+ frustum[2][2]*frustum[2][2]);
 	frustum[2][0] /= t;
@@ -123,7 +123,7 @@ void FRUSTUM::
 	frustum[3][3] = clip[15] - clip[13];
 	
 	/* normalize the top plane */
-	t = sqrt(frustum[3][0]*frustum[3][0] 
+	t = sqrtf(frustum[3][0]*frustum[3][0] 
 		+ frustum[3][1]*frustum[3][1] 
 		+ frustum[3][2]*frustum[3][2]);
 	frustum[3][0] /= t;
@@ -138,7 +138,7 @@ void FRUSTUM::
 	frustum[4][3] = clip[15] - clip[14];
 	
 	/* normalize the far plane */
-	t = sqrt(frustum[4][0]*frustum[4][0] 
+	t = sqrtf(frustum[4][0]*frustum[4][0] 
 		+ frustum[4][1]*frustum[4][1] 
 		+ frustum[4][2]*frustum[4][2]);
 	frustum[4][0] /= t;
@@ -153,7 +153,7 @@ void FRUSTUM::
 	frustum[5][3] = clip[15] + clip[14];
 	
 	/* normalize the near plane */
-	t = sqrt(frustum[5][0]*frustum[5][0] 
+	t = sqrtf(frustum[5][0]*frustum[5][0] 
 		+ frustum[5][1]*frustum[5][1] 
 		+ frustum[5][2]*frustum[5][2]);
 	frustum[5][0] /= t;
